initial refactor commit
