+++ social_app/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class ComingsoonConfig(AppConfig):
-    name = 'MainApp'
+    name = 'social_app'
